+++ src/SonnetDB.Core/Properties/AssemblyInfo.cs
@@ -2,3 +2,4 @@
 
 [assembly: InternalsVisibleTo("SonnetDB.Core.Tests")]
 [assembly: InternalsVisibleTo("SonnetDB.Benchmarks")]
+[assembly: InternalsVisibleTo("SonnetDB")]